@@ -2,20 +2,19 @@ name: in_app_purchases_paywall_ui
 description: >
   A collection of in app purchase Paywall as a UI library.
   Also handles Purchase success state. Currently only subscriptions are supported. Made by www.linkfive.io
-version: 0.6.0
+version: 0.7.0
 homepage: https://github.com/LinkFive/flutter-in_app_purchases_paywall
 repository: https://github.com/LinkFive/flutter-in_app_purchases_paywall
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ^3.0.0
+  flutter: '>=3.10.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_inappwebview: ^5.4.3+8
-  in_app_purchases_interface:
-    ^0.0.8
+  flutter_inappwebview: ^6.0.0 
+  in_app_purchases_interface: ^1.0.0
     #path: ../in_app_purchases_interface
 
 dev_dependencies:
